+++ packages/expo-router/src/fork/NavigationContainer.tsx
@@ -1,7 +1,6 @@
 import React from 'react';
 import { I18nManager } from 'react-native';
 
-import { useImperativeApiEmitter } from '../imperative-api';
 import type {
   DocumentTitleOptions,
   LinkingOptions,
@@ -83,7 +82,6 @@ function NavigationContainerInner(
 
   useBackButton(refContainer);
   useDocumentTitle(refContainer, documentTitle);
-  useImperativeApiEmitter(refContainer);
 
   const [lastUnhandledLink, setLastUnhandledLink] = React.useState<string | undefined>();