+++ packages/expo-router/src/global-state/store.ts
@@ -1,30 +1,19 @@
-import type { ComponentType } from 'react';
-
 import type { RouteNode } from '../Route';
 import type { ExpoLinkingOptions } from '../getLinkingConfig';
-import { resolveHref, resolveHrefStringWithSegments } from '../link/href';
 import type { NavigationContainerRefWithCurrent } from '../react-navigation/native';
-import type { Href } from '../types';
 import * as SplashScreen from '../views/Splash';
 import { defaultRouteInfo, type UrlObject } from './getRouteInfoFromState';
 import { getCachedRouteInfo, routeInfoSubscribers } from './routeInfoCache';
-import type {
-  FocusedRouteState,
-  LinkToOptions,
-  ReactNavigationState,
-  StoreRedirects,
-} from './types';
+import type { FocusedRouteState, ReactNavigationState } from './types';
 
 export type RouterStore = typeof store;
 
 type StoreRef = {
   navigationRef: NavigationContainerRefWithCurrent<ReactNavigation.RootParamList>;
   routeNode: RouteNode | null;
-  rootComponent: ComponentType<any>;
   state?: ReactNavigationState;
   linking?: ExpoLinkingOptions;
   config: any;
-  redirects: StoreRedirects[];
   routeInfo?: UrlObject;
 };
 
@@ -47,10 +36,18 @@ export function setHasAttemptedToHideSplash(value: boolean) {
   hasAttemptedToHideSplash = value;
 }
 
+export function maybeHideSplashScreen() {
+  if (!hasAttemptedToHideSplash) {
+    setHasAttemptedToHideSplash(true);
+    setSplashScreenAnimationFrame(
+      requestAnimationFrame(() => {
+        SplashScreen._internal_maybeHideAsync?.();
+      })
+    );
+  }
+}
+
 export const store = {
-  shouldShowTutorial() {
-    return !storeRef.current.routeNode && process.env.NODE_ENV === 'development';
-  },
   get state() {
     return storeRef.current.state;
   },
@@ -63,36 +60,12 @@ export const store = {
   getRouteInfo(): UrlObject {
     return storeRef.current.routeInfo || defaultRouteInfo;
   },
-  get redirects() {
-    return storeRef.current.redirects || [];
-  },
-  get rootComponent() {
-    return storeRef.current.rootComponent;
-  },
-  getStateForHref(href: Href | string, options?: LinkToOptions) {
-    href = resolveHref(href);
-
-    href = resolveHrefStringWithSegments(href, store.getRouteInfo(), options);
-    return this.linking?.getStateFromPath!(href, this.linking.config);
-  },
   get linking() {
     return storeRef.current.linking;
   },
   setFocusedState(state: FocusedRouteState) {
     const routeInfo = getCachedRouteInfo(state);
     storeRef.current.routeInfo = routeInfo;
-  },
-  // TODO(@ubax): Refactor onReady logic as it probably should live somewhere else then store
-  onReady() {
-    if (!hasAttemptedToHideSplash) {
-      setHasAttemptedToHideSplash(true);
-      // NOTE(EvanBacon): `navigationRef.isReady` is sometimes not true when state is called initially.
-      setSplashScreenAnimationFrame(
-        requestAnimationFrame(() => {
-          SplashScreen._internal_maybeHideAsync?.();
-        })
-      );
-    }
   },
   onStateChange(newState: ReactNavigationState | undefined) {
     if (!newState) {
@@ -125,13 +98,6 @@ export const store = {
 
     for (const callback of routeInfoSubscribers) {
       callback();
-    }
-  },
-  assertIsReady() {
-    if (!storeRef.current.navigationRef.isReady()) {
-      throw new Error(
-        'Attempted to navigate before mounting the Root Layout component. Ensure the Root Layout component is rendering a Slot, or other navigator on the first render.'
-      );
     }
   },
 };