+++ packages/expo-router/src/link/preview/__tests__/utils.test.ios.tsx
@@ -1,3 +1,4 @@
+import { store } from '../../../global-state/router-store';
 import { Stack } from '../../../layouts/Stack';
 import { NativeTabs } from '../../../native-tabs/index';
 import type { NavigationState } from '../../../react-navigation/native';
@@ -175,7 +176,12 @@ describe(getTabPathFromRootStateByHref, () => {
       ],
     };
     const href = '/faces/1e3a8a';
-    const tabPath = getTabPathFromRootStateByHref(href, state as NavigationState);
+    const tabPath = getTabPathFromRootStateByHref(
+      href,
+      state as NavigationState,
+      store.getRouteInfo(),
+      store.linking
+    );
     expect(tabPath).toEqual([
       {
         oldTabKey: 'faces-BlzNnnAhZ7c9t5bfSf4kR',
@@ -247,7 +253,12 @@ describe(getTabPathFromRootStateByHref, () => {
       ],
     };
     const href = '/faces/1e3a8a';
-    const tabPath = getTabPathFromRootStateByHref(href, state as NavigationState);
+    const tabPath = getTabPathFromRootStateByHref(
+      href,
+      state as NavigationState,
+      store.getRouteInfo(),
+      store.linking
+    );
     expect(tabPath).toEqual([
       {
         oldTabKey: 'index-rYeU6j6cRmkJK1pXpEFHs',
@@ -342,7 +353,12 @@ describe(getPreloadedRouteFromRootStateByHref, () => {
       ],
     };
     const href = '/faces/1e3a8a';
-    const preloadedRoute = getPreloadedRouteFromRootStateByHref(href, state as NavigationState);
+    const preloadedRoute = getPreloadedRouteFromRootStateByHref(
+      href,
+      state as NavigationState,
+      store.getRouteInfo(),
+      store.linking
+    );
     expect(preloadedRoute).toEqual({
       key: '[face]-9rms2gdsibY9dVYUGCpZG',
       name: '[face]',
@@ -415,7 +431,12 @@ describe(getPreloadedRouteFromRootStateByHref, () => {
       ],
     };
     const href = '/faces/1e3a8a';
-    const preloadedRoute = getPreloadedRouteFromRootStateByHref(href, state as NavigationState);
+    const preloadedRoute = getPreloadedRouteFromRootStateByHref(
+      href,
+      state as NavigationState,
+      store.getRouteInfo(),
+      store.linking
+    );
     expect(preloadedRoute).toEqual({
       key: '[face]-MZ5nYkDCFxwNv1BcD5exf',
       name: '[face]',